backend/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.0
flask>=2.0.0
flask-socketio>=5.0.0
pillow>=8.0.0
python-socketio>=5.0.0
eventlet>=0.31.0
twilio>=8.0.0
python-dotenv>=1.0.0